The offset for this decl should be 64-bit.

+++ include/vcblockchain/protocol.h
@@ -663,7 +663,7 @@ int vcblockchain_protocol_sendreq_extended_api(
  */
 int vcblockchain_protocol_sendreq_extended_api_response(
     ssock* sock, vccrypt_suite_options_t* suite, uint64_t* client_iv,
-    vccrypt_buffer_t* shared_secret, uint32_t offset,
+    vccrypt_buffer_t* shared_secret, uint64_t offset,
     const vccrypt_buffer_t* response_body);
 
 /**
